Rename file {specialist_document, section}_header_extractor
