+++ lib/section_header_extractor.rb
